pass db name to uploading section

+++ cads-edge/include/coms.h
@@ -23,7 +23,7 @@ namespace cads
   void remote_control_thread(moodycamel::BlockingConcurrentQueue<int> &,bool&);
   void publish_meta_realtime(std::string Id, double value, bool valid);
   void http_post_realtime(double y_area, double value);
-  std::tuple<date::utc_clock::time_point,bool> http_post_whole_belt(int, int, double);
+  std::tuple<date::utc_clock::time_point,bool> http_post_whole_belt(int, int, int);
   std::vector<profile> http_get_frame(double y, int len, date::utc_clock::time_point chrono);
   std::tuple<int,bool> remote_addconveyor(Conveyor params); 
 }
